CRUD in building service

+++ src/main/java/ua/edu/ukma/event_management_system/service/interfaces/EventService.java
@@ -6,9 +6,9 @@ public interface EventService {
 
     void createEvent(Event event);
 
-    Event getEventById(int eventId);
+    Event getEventById(long eventId);
 
     void updateEvent(Event event);
 
-    void deleteEvent(int eventId);
+    void deleteEvent(long eventId);
 }
